+++ raiden-dapp/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 ### Added
+- [#1222] Enables state upload
 - [#694] Show protocol progress steps on transfer modal.
 - [#687] Ghost action buttons.
 
@@ -10,6 +11,7 @@
 - [#1243] Fix width of transfer input on mobile screens.
 - [#1242] Fix transfer modal not closing instantly.
 
+[#1222] https://github.com/raiden-network/light-client/issues/1222
 [#694] https://github.com/raiden-network/light-client/issues/694
 [#1243]: https://github.com/raiden-network/light-client/issues/1243
 [#1242]: https://github.com/raiden-network/light-client/issues/1242